Move sortFrameworks func close to component
